tier/libgfdb : Setting Freq counters of un-selected files to zero
Change Time Recorder increments the write/read frequency counters
on a read or write of a file, if the "features.record-counters" is
"on". It is the responsibility of the tiering migrator to reset
these counters to zero for un-selected files to reset them to zero
as frequency counters are function of promotion/Demotion cycles.
If the counters are not set to zero then,

1) the counters may overflow in the DB
2) The file may be wrongly promoted or demoted.

This fix will reset the freq counters of un-selected files to zero
after promotion/demotion frequency.

Change-Id: Ideea2c76a52d421a7e67c37fb0c823f552b3da7a
BUG: 1242504
Signed-off-by: Joseph Fernandes <josferna@redhat.com>
Reviewed-on: http://review.gluster.org/11648
Tested-by: Joseph Fernandes
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Reviewed-by: Dan Lambright <dlambrig@redhat.com>