dht: send lookup even for fd based operations during rebalance
Problem:
dht_rebalance_inprogress_task() was not sending lookups to the
destination subvolume for a file undergoing writes during rebalance. Due to
this, afr was not able to populate the read_subvol and failed the write
with EIO.

Fix:
Send lookup for fd based operations as well.

Thanks to Raghavendra G for helping with the RCA.

Change-Id: I638c203abfaa45b29aa5902ffd76e692a8212a19
BUG: 1244165
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/11713
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: N Balachandran <nbalacha@redhat.com>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>