dht: send lookup even for fd based operations during rebalance
Backport of http://review.gluster.org/11713

Problem:
dht_rebalance_inprogress_task() was not sending lookups to the
destination subvolume for a file undergoing writes during rebalance. Due to
this, afr was not able to populate the read_subvol and failed the write
with EIO.

Fix:
Send lookup for fd based operations as well.

Thanks to Raghavendra G for helping with the RCA.

Change-Id: Iaa427666328109bbdf228876e62c13b75b7df88e
BUG: 1245934
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/11744
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: N Balachandran <nbalacha@redhat.com>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>