features/shard: Fix block size get from xdata
        Backport of: http://review.gluster.org/11778

Instead of using dict_get_ptr, dict_get_uint64 was used. If the first byte of
the value is '\0' then size is returned as 0 because strtoull is used in
data_to_uint64. This will make it seem like the file is not sharded at all.

Original author: Pranith Kumar K <pkarampu@redhat.com>
Change-Id: Id07a7d9523cb29d096b65dd68bbfcef395031aef
BUG: 1247833
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>
Signed-off-by: Krutika Dhananjay <kdhananj@redhat.com>
Reviewed-on: http://review.gluster.org/11789
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>