dht/tier :rename fails with EBUSY
When the files was in hot tier and the look up was done already, then
hashed and cached subvolume will be hot-tier. Once the file is moved
from hot-tier to cold-tier, then subsequent lookup will send a
revalidate lookup to hot-tier and it will find out that the file was
actually moved and there is only link in the cached subvolume. So dht
will return an ESTALE to fuse. Upon receiving ESTALE for a lookup, fuse
will create a new inode and sent a fresh lookup. This lookup will be
successful, and it will locate the file properly. Then fuse try to link
the inode, but the older inode was already there in inmemory inode cache
with same gfid and that is also shared with fuse kernal. So inode_link
will return the older ionode itself. So the subsequent rename fop will
come to gluster with the older inode. From dht_rename, we will take a
lock on the inode and after successful inodelk on inode dht will send
lookup before creating a link. this lookup will again find out that the
file is a link file, and then dht will think that file is
migrating/migrated in the mean time, and will send EBUSY.

Back port of :
>Change-Id: Ib3a01e5b1d7f64514b04bb6234026d049f082679
>BUG: 1248306
>Signed-off-by: Mohammed Rafi KC <rkavunga@redhat.com>
>Reviewed-on: http://review.gluster.org/11768
>Tested-by: Gluster Build System <jenkins@build.gluster.com>
>Reviewed-by: Raghavendra G <rgowdapp@redhat.com>
>Reviewed-by: Dan Lambright <dlambrig@redhat.com>
>Tested-by: NetBSD Build System <jenkins@build.gluster.org>
>Tested-by: Dan Lambright <dlambrig@redhat.com>

(cherry picked from commit 0ad26041fbf65ab36856a0ad178c32e51bf87319)

Change-Id: I1278a2c2ccc2cadcbe147db836f0526f079f6038
BUG: 1254437
Signed-off-by: Mohammed Rafi KC <rkavunga@redhat.com>
Reviewed-on: http://review.gluster.org/11944
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Dan Lambright <dlambrig@redhat.com>
Tested-by: Dan Lambright <dlambrig@redhat.com>