features/shard: Fill inode ctx in readdir(p) callback too
        Backport of: http://review.gluster.org/#/c/11854/

The only place where shard translator was initialising inode ctx
was lookup callback. But if the inodes are created and linked through
readdirp, shard_lookup() path _may_ not be exercised before FUSE
winds other fops on them. Since shard translator does an
inode_ctx_get() first thing in most fops, an uninitialised ctx could
cause it to fail the operation with ENOMEM.

The solution would be to also initialise inode ctx if it has not been
done already in readdir(p) callback.

Change-Id: I6384e2d4c6b443c02c6620bbbc38279ddfad18ae
BUG: 1251106
Signed-off-by: Krutika Dhananjay <kdhananj@redhat.com>
Reviewed-on: http://review.gluster.org/11866
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>