afr: Do not wind statfs to arbiter brick
Backport of: http://review.gluster.org/#/c/11857/

Problem: AFR serves statfs from the brick having the least free space
available. Since the size to be allocated to the arbiter brick in a 3
way replica is supposed to be considerably lesser than the other 2
bricks, statfs will be served from this brick which is incorrect.

Fix: Don't serve statfs from the arbiter brick.

Change-Id: Ia2d2402ba1e8f5d96831f71b3f8337f241e6753b
BUG: 1251380
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/11858
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>