afr: launch index heal on local subvols up on a child-up event
Backport of http://review.gluster.org/#/c/11912/
Problem:
When a replica's child goes down and comes up, the index heal is
triggered only on the child that just came up. This does not serve the
intended purpose as the list of files that need to be healed
to this child is actually captured on the other child of the replica.

Fix:
Launch index-heal on all local children of the replica xlator which just
received a child up. Note that afr_selfheal_childup() eventually calls
afr_shd_index_healer() which will not run the heal on non-local
children.

Signed-off-by: Ravishankar N <ravishankar@redhat.com>

Change-Id: Ia23e47d197f983c695ec0bcd283e74931119ee55
BUG: 1255690
Reviewed-on: http://review.gluster.org/11982
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Krutika Dhananjay <kdhananj@redhat.com>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>