afr: launch index heal on local subvols up on a child-up event
Backport of http://review.gluster.org/#/c/11912/

Problem: When a replica's child goes down and comes up, the index heal is
triggered only on the child that just came up. This does not serve the intended
purpose as the list of files that need to be healed to this child is actually
captured on the other child of the replica.

Fix: Launch index-heal on all local
children of the replica xlator which just received a child up. Note that
afr_selfheal_childup() eventually calls afr_shd_index_healer() which will not
run the heal on non-local children.

Change-Id: I524fda17c28864758b35679cfb232f81f8374571
BUG: 1256245
Signed-off-by: Ravishankar N <ravishankar@redhat.com>
Reviewed-on: http://review.gluster.org/11994
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Reviewed-by: Raghavendra Bhat <raghavendra@redhat.com>
Tested-by: Raghavendra Bhat <raghavendra@redhat.com>