posix: xattrop 'GF_XATTROP_GET_AND_SET' implementation
This is a backport of http://review.gluster.org/#/c/11995

GF_XATTROP_GET_AND_SET stores the existing xattr
value in xdata and sets the new value

xattrop was reusing input xattr dict to set the results
instead of creating new dict.
This can be problem for server side xlators as the inout dict
will have the value changed.

> Change-Id: I43369082e1d0090d211381181e9f3b9075b8e771
> BUG: 1251454
> Signed-off-by: vmallika <vmallika@redhat.com>

Change-Id: I7e0c27fd415131e9983a10d27067f63ed3a7701e
BUG: 1257441
Signed-off-by: vmallika <vmallika@redhat.com>
Reviewed-on: http://review.gluster.org/12022
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Reviewed-by: Raghavendra G <rgowdapp@redhat.com>