features/shard: Fix unlink failure due to non-existent shard(s)
        Backport of: http://review.gluster.org/#/c/12059/

Unlink of a sharded file with holes was leading to EINVAL errors
because it was being wound on non-existent shards (those blocks that
fall in the hole region). loc->inode was NULL in these cases and
dht_unlink used to fail the FOP with EINVAL for failure to fetch
cached subvol for the inode.

The fix involves winding unlink on only those shards whose corresponding
inodes exist in memory.

Change-Id: I1e5d492a2e60491601da23f64a5d0089e536b305
BUG: 1258353
Signed-off-by: Krutika Dhananjay <kdhananj@redhat.com>
Reviewed-on: http://review.gluster.org/12061
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>