tests: add bug-948686.t to bad_tests
bug-948686.t intermittely ends up in a race in GlusterD where one thread deletes
a stale volume and on a volume start request a start is attempted on the same
object which might be freed up by the other thread. We are considering to use
refcounting mechanism while accessing volinfo objects to get rid of this problem
but that needs a significant amount of time in development. Till then adding
this test in bad_tests ()

Change-Id: I3ddc025b07c47b17ea75554d5cc809a839365f10
BUG: 1258714
Signed-off-by: Atin Mukherjee <amukherj@redhat.com>
Reviewed-on: http://review.gluster.org/12070
Reviewed-by: Raghavendra Talur <rtalur@redhat.com>
Tested-by: NetBSD Build System <jenkins@build.gluster.org>
Tested-by: Gluster Build System <jenkins@build.gluster.com>